+++ routers/index.js
@@ -45,6 +45,9 @@ module.exports = _ => {
   _.get("/workflow/list", authentication, workflow.list); // 工作流信息列表
   _.get("/workflow/:id/simple", authentication, workflow.simpleInfo); // 获取工作流实例简单信息
   _.get("/workflow/:nodeId/files", authentication, workflow.nodeFiles); // 获取工作流某实例下的文件列表
+  _.get("/workflow/overviews/platform", authentication, workflow.overviewsPlatform); // 总览 - 平台工作流
+  _.get("/workflow/overviews/product", authentication, workflow.overviewsProduct);// 总览 - 产品工作流
+  _.get("/workflow/overviews/directory", authentication, workflow.overviewsDirectory); // 总览 - 公共目录
 
   /*
    *  文件系统模块